fix: Resolve PHPStan errors in Compliance domain
- Fix variable name typo in KycSubmissionActivity
- Add getter methods to ComplianceAggregate to resolve property-only-written warnings
- Cast return types in SuspiciousActivityReport methods
- Remove invalid country property from User factory in tests
- Add PHPStan ignore for intentional readonly property test

+++ app/Domain/Compliance/Models/SuspiciousActivityReport.php
@@ -388,7 +388,7 @@ class SuspiciousActivityReport extends Model
     {
         // This would typically calculate from related transactions
         // For now, return the stored total_amount
-        return $this->total_amount ?? 0;
+        return (float) ($this->total_amount ?? 0);
     }
 
     public function getActivityDuration(): int
@@ -397,7 +397,7 @@ class SuspiciousActivityReport extends Model
             return 0;
         }
 
-        return $this->activity_start_date->diffInDays($this->activity_end_date);
+        return (int) $this->activity_start_date->diffInDays($this->activity_end_date);
     }
 
     public function getSummary(): array
